move sploit.c out to data folder
